Class "dimRedMethod"

A virtual class "dimRedMethod" to serve as a template to implement methods for dimensionality reduction.

Details

Implementations of dimensionality reductions should inherit from this class.

The fun slot should be a function that takes three arguments

data

An object of class dimRedData.

pars

A list with the standard parameters.

keep.org.data

Logical. If the original data should be kept in the output.

and returns an object of class dimRedResult.

The stdpars slot should take a list that contains standard parameters for the implemented methods.

This way the method can be called by embed(data, "method-name", ...), where ... can be used to to change single parameters.

Slots

fun

A function that does the embedding.

stdpars

A list with the default parameters for the fun slot.

requires

A vector with all packages R packages that need to be installed to run the method. In some occasions a method may work without one of the packages. Does not include Python dependencies such as Tensorflow. Used to auto skip tests

See also