dimRedMethodList

Get the names of all methods for dimensionality reduction.

Usage

dimRedMethodList(filter = FALSE)

Arguments

filter

filter methods by methods that have their dependencies installed

Value

a character vector with the names of classes that inherit from dimRedMethod.

Details

Returns the name of all classes that inherit from dimRedMethod-class to use with embed.

See also

Examples

dimRedMethodList()
#>  [1] "DiffusionMaps"       "DRR"                 "FastICA"            
#>  [4] "KamadaKawai"         "DrL"                 "FruchtermanReingold"
#>  [7] "HLLE"                "Isomap"              "kPCA"               
#> [10] "PCA_L1"              "MDS"                 "nMDS"               
#> [13] "NNMF"                "PCA"                 "tSNE"               
#> [16] "UMAP"