dimRedMethodList

Get the names of all methods for dimensionality reduction.

Usage

dimRedMethodList(filter = FALSE)

Arguments

filter

filter methods by methods that have their dependencies installed

Value

a character vector with the names of classes that inherit from dimRedMethod.

Details

Returns the name of all classes that inherit from dimRedMethod-class to use with embed.

See also

Examples

dimRedMethodList()
#>  [1] "AutoEncoder"         "DiffusionMaps"       "DRR"                
#>  [4] "FastICA"             "KamadaKawai"         "DrL"                
#>  [7] "FruchtermanReingold" "HLLE"                "Isomap"             
#> [10] "kPCA"                "PCA_L1"              "LaplacianEigenmaps" 
#> [13] "MDS"                 "nMDS"                "NNMF"               
#> [16] "PCA"                 "tSNE"                "UMAP"